Wave 25 of the 2014-2023 British Election Study Internet Panel

The survey dataset is from the British Electoral Study. It is purely as a way to practice using the functions and charts within this package. The fieldwork for the survey was conducted by YouGov between 5 May and 23 May, 2023. The practice dataset has n=5000 respondents and 65 variables. The original dataset has 570 variables and n=30,407 respondents with 63% retained from Wave 23 (May 2022) and 74% from Wave 24 (December 2022). The dataset includes a cumulative weight (wt).

Usage

survey

Format

A data frame with 5,000 rows and 65 variables:

id

unique identifier of respondent

wt

weight variable

turnoutUKGeneral

Likelihood to vote in general election

generalElectionVote

General election vote intention (recalled vote in post-election waves)

partyIdStrength

Strength of party identification

partyId

Party identification

partyIdSqueeze

Party ID if no party to first ID question

mii

Most Important Issue

mii_cat

MII manual coding

small_mii_cat

MII (manually-coded-collapsed)

LRAL_mii_cat

Dimension of MII

bestOnMII

Best party on most important issue

polAttention

Attention to Politics

pidWeThey

When I speak about this party, I usually say "we" instead of "they".

pidInterestedOthers

I am interested in what other people think about this party.

pidCriticiseParty

When people criticize this party, it feels like a personal insult.

pidCommonParty

I have a lot in common with other supporters of this party.

pidConnected

When I meet someone who supports this party, I feel connected with this person.

pidPraiseGood

When people praise this party, it makes me feel good.

pidWeTheyb

When I speak about this party, I usually say "we" instead of "they".

pidInterestedOthersb

I am interested in what other people think about this party.

pidCriticisePartyb

When people criticize this party, it feels like a personal insult.

pidCommonPartyb

I have a lot in common with other supporters of this party.

pidConnectedb

When I meet someone who supports this party, I feel connected with this person.

pidPraiseGoodb

When people praise this party, it makes me feel good.

likeSunak

Like-dislike: Rishi Sunak

likeStarmer

Like-dislike: Keir Starmer

likeCon

Like-dislike: Conservatives

likeLab

Like-dislike: Labour

likeLD

Like-dislike: Liberal Democrats

likeSNP

Like-dislike: Scottish National Party

likePC

Like-dislike: Plaid Cymru

likeBrexitParty

Like-dislike: Brexit Party

likeGrn

Like-dislike: Green Party

country

Country

gender

Gender

age

Age

ageGroup

Age group

pcon

Parliamentary Constituency

p_education

Education qualification (highest attained)

p_work_stat

Which of these applies to you?

p_hh_children

How many of the people in your household are under 18?

p_housing

Do you own or rent the home in which you live?

p_gross_household

Gross household income

p_gross_personal

Gross personal income

p_hh_size

How many people, including yourself, are there in your household? Please includ

p_socgrade

Social Grade

p_disability

Are your day-to-day activities limited because of a health problem or disabilit

p_sexuality

Which of the following best describes your sexuality?

p_job_sector

What kind of organisation do you work for?

p_education_age

Age completed formal education

p_marital

What is your current marital or relationship status?

p_paper_read

Which daily newspaper do you read most often?

p_religion

Do you regard yourself as belonging to any particular religion, and if so, to w

p_parent

Are you a parent or guardian?

p_country_birth

Country of birth

p_ethnicity

To which of these groups do you consider you belong?

p_past_vote_2010

2010 GE vote choice

p_past_vote_2005

2005 GE vote choice

p_past_vote_2015

2015 GE vote choice

p_past_vote_2017

2017 GE vote choice

p_past_vote_2019

2019 GE vote choice

p_eurefvote

EU referendum vote (earliest recorded)

p_edlevel

Education level

p_eurefturnout

EU referendum turnout (earliest recorded)

Source

This dataset was sourced from the British Electoral Study website (Wave 25). The raw data are available from https://www.britishelectionstudy.com/data-object/wave-25-of-the-2014-2023-british-election-study-internet-panel/.

Details

A full description of themethodology and questionnaire can be found here: https://www.britishelectionstudy.com/wp-content/uploads/2023/11/Bes_wave25Documentation.pdf