Sankey Diagram Visualisation

plot_sankey enhances networkD3::sankeyNetwork and htmlwidgets to create a Sankey diagram, a specific type of flow diagram that visually represents the movement or flow of data between different entities (referred to as "nodes"). It is particularly useful for illustrating the distribution of votes flowing between parties across elections or in preferential voting systems.

Usage

plot_sankey(
  data,
  source,
  target,
  value,
  units = "",
  colours = colour_pal("catExtended"),
  fontSize = 20,
  fontFamily = "Calibri",
  nodeWidth = 20,
  nodePadding = 10,
  margin = list(left = 0, right = 0),
  width = 1200,
  height = 800,
  shiftLabel = NA
)

Arguments

data

A data frame containing the flow data, where each row represents a unique flow from a source to a target.

source

The column in the data frame representing the source node of the flow.

target

The column in the data frame representing the target node of the flow.

value

The column in the data frame representing the magnitude or volume of the flow between the source and target.

units

A string indicating the units of the 'value'. This is displayed in the tooltip of the diagram. The default is an empty string, indicating that no units are specified.

colours

A list or vector of colours (in hexadecimal format or colour names) to be used for the nodes and links in the diagram. If a named list is provided, the names should correspond to the node names for specific colour mapping.

fontSize

The font size for text in the diagram. Default is set to 20.

fontFamily

The font family for text in the diagram. Default is "Calibri".

nodeWidth

The width of the nodes (entities) in the diagram. Default is 20.

nodePadding

The padding between nodes in the diagram. Default is 10.

margin

A list specifying the margins around the diagram. Default is list("left" = 0, "right" = 0). Increase the margin if the node names are long and go off the page.

width

The width of the plot in pixels. Default is 1200.

height

The height of the plot in pixels. Default is 800.

shiftLabel

A numeric value controlling the positioning of labels on the nodes. When set to NA, labels are automatically placed outside the nodes. If the automatic placement results in uneven label distribution, this parameter can be adjusted to achieve the desired layout.

Value

A Sankey diagram visualising the flow of data as specified by the parameters. The diagram is interactive, allowing for the exploration of individual flows and nodes.

Examples

if (FALSE) {
# Prepare data
split <- grp_freq(df,
                  c("Q1a","Q1c"),
                  "wgtvar",
                  c("Party_Vote","Candidate_Vote","Votes")) %>%
         mutate(Votes = Votes/sum(Votes))
#           Party_Vote    Candidate_Vote        Votes
# 1       Labour Party      Labour Party 0.1036125880
# 2     National Party      Labour Party 0.1040036893
# 3        Green Party      Labour Party 0.0320547032
# ...

# Create sankey
plot_sankey(data = split,
            source = "Party_Vote",
            target = "Candidate_Vote",
            value = "Votes",
            colours = colour_prep(df, c("Q1a", "Q1c"), pal_name = "polNZ"))
}